Tried to implement some Utility::Bitconverter methods but it didnt worked InternalTags cant be higher than 255 because they will be converted to byte added new ToString method in Utility::StringConverter Server now closes the connection better

+++ src/TcpClientWindows.cpp
@@ -75,8 +75,8 @@ void TcpClient::Shutdown()
 {
 	NetworkMessage message;
 	message.sender = id;
-	message.distribution_mode = Server;
-	message.tag = DISCONNECT;
+	message.distribution_mode = static_cast<byte>(Server);
+	message.tag = static_cast<byte>(DISCONNECT);
 	SendMessage(message);
 	uint16 code = closesocket(tcp_socket);
 	if (code == SOCKET_ERROR)
